+++ scripts/breaking_changes_checker/tests/test_changelog.py
@@ -8,7 +8,7 @@
 import os
 import json
 import jsondiff
-from breaking_changes_checker.breaking_changes_tracker import BreakingChangesTracker, ChangeType
+from breaking_changes_checker.changelog_tracker import ChangelogTracker
 
 
 def test_changelog_flag():
@@ -18,12 +18,12 @@ def test_changelog_flag():
         current = json.load(fd)
     diff = jsondiff.diff(stable, current)
 
-    bc = BreakingChangesTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
+    bc = ChangelogTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
     bc.run_checks()
 
     assert len(bc.features_added) > 0
     msg, _, *args = bc.features_added[0]
-    assert msg == BreakingChangesTracker.ADDED_CLIENT_METHOD_MSG
+    assert msg == ChangelogTracker.ADDED_CLIENT_METHOD_MSG
     assert args == ['azure.ai.contentsafety', 'BlocklistClient', 'new_blocklist_client_method']
 
 def test_new_class_property_added():
@@ -58,12 +58,12 @@ def test_new_class_property_added():
     }
 
     diff = jsondiff.diff(stable, current)
-    bc = BreakingChangesTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
+    bc = ChangelogTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
     bc.run_checks()
 
     assert len(bc.features_added) == 1
     msg, _, *args = bc.features_added[0]
-    assert msg == BreakingChangesTracker.ADDED_CLASS_PROPERTY_MSG
+    assert msg == ChangelogTracker.ADDED_CLASS_PROPERTY_MSG
     assert args == ['azure.ai.contentsafety', 'AnalyzeTextResult', 'new_class_att']
 
 
@@ -135,12 +135,12 @@ def test_new_class_property_added_init():
     }
 
     diff = jsondiff.diff(stable, current)
-    bc = BreakingChangesTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
+    bc = ChangelogTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
     bc.run_checks()
 
     assert len(bc.features_added) == 1
     msg, _, *args = bc.features_added[0]
-    assert msg == BreakingChangesTracker.ADDED_CLASS_PROPERTY_MSG
+    assert msg == ChangelogTracker.ADDED_CLASS_PROPERTY_MSG
     assert args == ['azure.ai.contentsafety', 'AnalyzeTextResult', 'new_class_att']
 
 
@@ -207,12 +207,12 @@ def test_new_class_property_added_init_only():
     }
 
     diff = jsondiff.diff(stable, current)
-    bc = BreakingChangesTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
+    bc = ChangelogTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
     bc.run_checks()
 
     assert len(bc.features_added) == 1
     msg, _, *args = bc.features_added[0]
-    assert msg == BreakingChangesTracker.ADDED_CLASS_METHOD_PARAMETER_MSG
+    assert msg == ChangelogTracker.ADDED_CLASS_METHOD_PARAMETER_MSG
     assert args == ['azure.ai.contentsafety', 'AnalyzeTextResult', 'new_class_att', '__init__']
 
 
@@ -296,10 +296,10 @@ def test_new_class_method_parameter_added():
     }
 
     diff = jsondiff.diff(stable, current)
-    bc = BreakingChangesTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
+    bc = ChangelogTracker(stable, current, diff, "azure-ai-contentsafety", changelog=True)
     bc.run_checks()
 
     assert len(bc.features_added) == 1
     msg, _, *args = bc.features_added[0]
-    assert msg == BreakingChangesTracker.ADDED_CLASS_METHOD_PARAMETER_MSG
+    assert msg == ChangelogTracker.ADDED_CLASS_METHOD_PARAMETER_MSG
     assert args == ['azure.ai.contentsafety', 'AnalyzeTextResult', 'bar', 'foo']
